refactor to a forms.py
